+++ providers/filesystem/filesystem_test.go
@@ -6,11 +6,15 @@ package filesystem
 import (
 	"bytes"
 	"context"
+	"os"
 	"strings"
 	"sync"
 	"testing"
+	"time"
 
 	"github.com/efficientgo/core/testutil"
+
+	"github.com/thanos-io/objstore"
 )
 
 func TestDelete_EmptyDirDeletionRaceCondition(t *testing.T) {
@@ -59,6 +63,60 @@ func TestIter_CancelledContext(t *testing.T) {
 
 	testutil.NotOk(t, err)
 	testutil.Equals(t, context.Canceled, err)
+}
+
+func TestIterWithAttributes(t *testing.T) {
+	dir := t.TempDir()
+	f, err := os.CreateTemp(dir, "test")
+	testutil.Ok(t, err)
+	defer f.Close()
+
+	stat, err := f.Stat()
+	testutil.Ok(t, err)
+
+	cases := []struct {
+		name              string
+		opts              []objstore.IterOption
+		expectedUpdatedAt time.Time
+	}{
+		{
+			name: "no options",
+			opts: nil,
+		},
+		{
+			name: "with updated at",
+			opts: []objstore.IterOption{
+				objstore.WithUpdatedAt(),
+			},
+			expectedUpdatedAt: stat.ModTime(),
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			b, err := NewBucket(dir)
+			testutil.Ok(t, err)
+
+			var attrs objstore.IterObjectAttributes
+
+			ctx := context.Background()
+			err = b.IterWithAttributes(ctx, "", func(objectAttrs objstore.IterObjectAttributes) error {
+				attrs = objectAttrs
+				return nil
+			}, tc.opts...)
+
+			testutil.Ok(t, err)
+
+			lastModified, ok := attrs.LastModified()
+			if zero := tc.expectedUpdatedAt.IsZero(); zero {
+				testutil.Equals(t, false, ok)
+			} else {
+				testutil.Equals(t, true, ok)
+				testutil.Equals(t, tc.expectedUpdatedAt, lastModified)
+			}
+		})
+
+	}
 }
 
 func TestGet_CancelledContext(t *testing.T) {